baseTaskManager/requirements.txt
PyMySQL
pyotp
qrcode
Flask
flask-login
Flask-SQLAlchemy
uwsgi
cryptography
Pillow
icecream

